requirements.txt
numpy==1.24.4
openai-clip==1.0.1
opencv-python==4.5.5.62
requests==2.28.1
segment-anything==1.0
tqdm==4.64.1
matplotlib==3.7.0
torch==2.0.1+cu117
torchvision==0.15.2+cu117